backend/app/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Email (for password reset)
python-dotenv==1.0.0
emails==0.6

# ML/Data Science (for predictions)
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.6.1
joblib==1.3.2

# Utilities
python-dateutil==2.8.2

